refactor: extract resolver to _internal module
