+++ src/conformly/_internal/generator/orchestration.py
@@ -4,9 +4,13 @@ from .context import GenerationContext
 from .registry import choose_mismatch_kind, get_generator
 
 from conformly._internal.planner import PlannedTask
+from conformly._internal.resolver import (
+    ResolvedField,
+    ResolvedModel,
+    create_minimal_semantic,
+)
+from conformly._internal.resolver.semantics import ListSemantic
 from conformly._internal.types import UNSET, FieldKind, FieldPath, ViolationType
-from conformly.resolver import ResolvedField, ResolvedModel, create_minimal_semantic
-from conformly.resolver.semantics import ListSemantic
 
 
 def generate_valid(ctx: GenerationContext, model: ResolvedModel) -> dict[str, Any]:
