+++ src/conformly/_internal/generator/types/enum.py
@@ -2,8 +2,8 @@ from typing import Any
 
 from ..context import GenerationContext
 
+from conformly._internal.resolver.semantics import EnumSemantic
 from conformly._internal.types import ViolationType
-from conformly.resolver.semantics import EnumSemantic
 
 _INVALID_ENUM_PREFIX = "__INVALID_ENUM_"
