+++ src/conformly/_internal/generator/types/ipv4.py
@@ -1,7 +1,7 @@
 from ..context import GenerationContext
 
+from conformly._internal.resolver.semantics import StringSemantic
 from conformly._internal.types import ViolationType
-from conformly.resolver.semantics import StringSemantic
 
 TEST_RANGES = [
     ("192.0.2", 1, 254),
