+++ src/conformly/_internal/generator/types/ipvany.py
@@ -1,8 +1,8 @@
 from ..context import GenerationContext
 from . import ipv4, ipv6
 
+from conformly._internal.resolver.semantics import StringSemantic
 from conformly._internal.types import ViolationType
-from conformly.resolver.semantics import StringSemantic
 
 
 def generate_value(
