+++ src/conformly/_internal/generator/types/string.py
@@ -3,8 +3,8 @@ import string
 
 from ..context import GenerationContext
 
+from conformly._internal.resolver.semantics.string import StringSemantic
 from conformly._internal.types import MAX_GENERATION_ATTEMPTS, ViolationType
-from conformly.resolver.semantics.string import StringSemantic
 
 DEFAULT_MIN_LENGTH = 5
 DEFAULT_MAX_LENGTH = 15
