+++ src/conformly/_internal/planner/field.py
@@ -1,13 +1,8 @@
 from .model import PlannedTask
 
 from conformly._internal.fields import SPECIAL_KINDS
-from conformly._internal.types import (
-    FieldKind,
-    FieldPath,
-    ViolationType,
-)
-from conformly.resolver import ResolvedModel
-from conformly.resolver.semantics import (
+from conformly._internal.resolver import ResolvedModel
+from conformly._internal.resolver.semantics import (
     BooleanSemantic,
     EnumSemantic,
     FieldSemantics,
@@ -16,6 +11,11 @@ from conformly.resolver.semantics import (
     ObjectSemantic,
     StringSemantic,
     UUIDSemantic,
+)
+from conformly._internal.types import (
+    FieldKind,
+    FieldPath,
+    ViolationType,
 )
 
 _VIOLATION_PRIORITY: tuple[ViolationType, ...] = (
