+++ src/conformly/_internal/resolver/__init__.py
@@ -1,5 +1,4 @@
-from .field import ResolvedField
-from .model import ResolvedModel
+from .models import ResolvedField, ResolvedModel
 from .resolve import resolve_model
 from .semantics import FieldSemantics
 from .semantics.factory import create_minimal_semantic
