+++ src/conformly/_internal/resolver/models.py
@@ -1,11 +1,37 @@
 from __future__ import annotations
 
 from dataclasses import dataclass, field
-from typing import TYPE_CHECKING
+from typing import TYPE_CHECKING, Any
 
 if TYPE_CHECKING:
-    from .._internal.types import FieldPath
-    from .field import ResolvedField
+    from .semantics import FieldSemantics
+
+    from conformly._internal.parser import FieldSpec
+    from conformly._internal.types import FieldPath
+
+
+@dataclass(frozen=True)
+class ResolvedField:
+    field_spec: FieldSpec
+    path: FieldPath
+    semantic: FieldSemantics
+    nested_model: ResolvedModel | None = None
+
+    @property
+    def name(self) -> str:
+        return self.field_spec.name
+
+    @property
+    def py_type(self) -> type:
+        return self.field_spec.field_type
+
+    @property
+    def default(self) -> Any:
+        return self.field_spec.default
+
+    @property
+    def nullable(self) -> bool:
+        return self.field_spec.nullable
 
 
 @dataclass(frozen=True)
