+++ src/conformly/_internal/resolver/semantics/uuid.py
@@ -1,6 +1,6 @@
 from dataclasses import dataclass
 
-from ..._internal.types import FieldKind
+from conformly._internal.types import FieldKind
 
 
 @dataclass(frozen=True)
