+++ tests/benchmarks/test_pydantic.py
@@ -7,7 +7,7 @@ from pydantic import BaseModel, Field
 
 from conformly import case, cases
 from conformly._internal.parser.adapters.pydantic import parse
-from conformly.resolver.resolve import resolve_model
+from conformly._internal.resolver.resolve import resolve_model
 
 
 class PydanticAddress(BaseModel):
