+++ tests/test_generator/test_types_generators/test_integer_generator.py
@@ -5,8 +5,8 @@ from conformly._internal.generator.types.integer import (
     _generate_invalid_integer,
     generate_value,
 )
+from conformly._internal.resolver.semantics.numeric import NumericSemantic
 from conformly._internal.types import INT_MAX, INT_MIN, FieldKind, Range, ViolationType
-from conformly.resolver.semantics.numeric import NumericSemantic
 
 int_semantic_10_20 = NumericSemantic(
     kind=FieldKind.INTEGER,
