+++ tests/test_generator/test_types_generators/test_ipv4_generator.py
@@ -7,8 +7,8 @@ from conformly._internal.generator.types.ipv4 import (
     _generate_valid_ipv4,
     generate_value,
 )
+from conformly._internal.resolver.semantics import StringSemantic
 from conformly._internal.types import FieldKind, LengthRange, ViolationType
-from conformly.resolver.semantics import StringSemantic
 
 
 def test_generate_valid_ipv4_format(ctx: GenerationContext) -> None:
