+++ tests/test_resolver/test_resolve.py
@@ -18,19 +18,8 @@ from conformly._internal.constraints import (
 )
 from conformly._internal.fields import SPECIAL_KINDS, Email
 from conformly._internal.parser import FieldSpec, ModelSpec
-from conformly._internal.types import (
-    ENUMERATED_TYPE,
-    FLOAT_MAX,
-    FLOAT_MIN,
-    INT_MAX,
-    INT_MIN,
-    FieldKind,
-    FieldPath,
-    LengthRange,
-    Range,
-)
-from conformly.resolver import ResolvedModel
-from conformly.resolver.resolve import (
+from conformly._internal.resolver import ResolvedModel
+from conformly._internal.resolver.resolve import (
     _build_indexes,
     _extract_numeric_multiple_of,
     calculate_invalid_numeric_ranges,
@@ -42,7 +31,7 @@ from conformly.resolver.resolve import (
     resolve_field,
     resolve_model,
 )
-from conformly.resolver.semantics import (
+from conformly._internal.resolver.semantics import (
     BooleanSemantic,
     EnumSemantic,
     ListSemantic,
@@ -50,7 +39,18 @@ from conformly.resolver.semantics import (
     ObjectSemantic,
     StringSemantic,
 )
-from conformly.resolver.semantics.uuid import UUIDSemantic
+from conformly._internal.resolver.semantics.uuid import UUIDSemantic
+from conformly._internal.types import (
+    ENUMERATED_TYPE,
+    FLOAT_MAX,
+    FLOAT_MIN,
+    INT_MAX,
+    INT_MIN,
+    FieldKind,
+    FieldPath,
+    LengthRange,
+    Range,
+)
 
 
 @pytest.fixture
